Use <img> for logo SVG to save 30 DOM elements count - Lighthouse wants maximum of 800 elements, about 48 more to go - Set img size width to avoid content layout shift
