+++ packages/scrabble-solver/next.config.js
@@ -35,8 +35,16 @@ module.exports = {
         ...config.module.rules,
         {
           test: /\.svg$/,
+          include: [path.resolve(__dirname, 'src/icons')],
           issuer: /\.tsx?$/,
           use: ['@svgr/webpack'],
+        },
+        {
+          test: /\.(png|jpg|jpeg|gif|webp|avif|ico|bmp|svg)$/i,
+          exclude: [path.resolve(__dirname, 'src/icons')],
+          loader: 'next-image-loader',
+          issuer: /\.tsx?$/,
+          options: { isServer: true, isDev: true, basePath: '', assetPrefix: '' },
         },
       ],
     },
