+++ packages/scrabble-solver/src/components/Logo/Logo.tsx
@@ -1,21 +1,7 @@
-/**
- * This component is unused, but it serves as a blueprint for the Logo.svg.
- * Logo.svg is what this component generates with all the text nodes transformed
- * into paths (manually with Inkscape), and corner radius removed from tiles.
- */
-import { CSSProperties, FunctionComponent } from 'react';
+import { FunctionComponent, HTMLProps } from 'react';
 
-import PlainTiles from '../PlainTiles';
+import logo from './Logo.svg';
 
-interface Props {
-  className?: string;
-  style?: CSSProperties;
-}
-
-const CONTENT = [['SCRABBLE'], ['SOLVER', '2']];
-
-const Logo: FunctionComponent<Props> = ({ className, style }) => (
-  <PlainTiles className={className} content={CONTENT} style={style} />
-);
+const Logo: FunctionComponent<HTMLProps<HTMLImageElement>> = (props) => <img {...props} src={logo.src} />;
 
 export default Logo;
