+++ packages/scrabble-solver/src/components/Logo/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Logo.svg';
+export { default } from './Logo';
