+++ packages/scrabble-solver/src/styles/variables.scss
@@ -108,6 +108,7 @@ $easeOutSine: cubic-bezier(0.61, 1, 0.88, 1);
 
   --button--icon--size: 24px;
   --dictionary--height: 260px;
+  --logo--aspect-ratio: 682 / 166;
   --logo--height: 60px;
   --modal--width: 370px;
   --nav--height: calc(var(--logo--height) + var(--nav--padding));
